src/core/auth-service.ts
import { createHash, randomBytes, timingSafeEqual } from "crypto";
import { Credentials, AuthResult, SecurityContext } from "../models/auth.model";

/**
 * Handles authentication and security-related operations for database connections
 */
export class AuthService {
  private readonly saltRounds = 10;
  private readonly sessionDuration = 3600; // 1 hour in seconds

  /**
   * Authenticates a client using provided credentials
   * Uses timing-safe comparison to prevent timing attacks
   */
  async authenticate(credentials: Credentials): Promise<AuthResult> {
    try {
      // Validate TLS/SSL first if certificate is provided
      if (credentials.clientCertificate) {
        const tlsValid = await this.validateTLSCertificate(
          credentials.clientCertificate
        );
        if (!tlsValid) {
          return this.createFailedResult("Invalid TLS certificate");
        }
      }

      // Perform credential validation
      // In production, this would check against a secure credential store
      const isValid = await this.validateCredentials(credentials);
      if (!isValid) {
        return this.createFailedResult("Invalid credentials");
      }

      // Create security context for the session
      const securityContext = this.createSecurityContext();

      return {
        success: true,
        sessionId: this.generateSessionId(),
        securityContext,
      };
    } catch (error) {
      return this.createFailedResult(
        `Authentication failed: ${
          error instanceof Error ? error.message : "Unknown error"
        }`
      );
    }
  }

  /**
   * Validates TLS/SSL certificate
   * In production, this would verify against trusted CAs
   */
  private async validateTLSCertificate(cert: string): Promise<boolean> {
    // Implementation would verify certificate chain, expiration, and revocation status
    return true; // Placeholder
  }

  /**
   * Validates user credentials using timing-safe comparison
   * Prevents timing attacks by ensuring constant-time comparison
   */
  private async validateCredentials(
    credentials: Credentials
  ): Promise<boolean> {
    // In production, this would:
    // 1. Retrieve hashed password from secure storage
    // 2. Use proper password hashing (e.g., bcrypt, Argon2)
    // 3. Perform timing-safe comparison

    const mockHashedPassword = await this.hashPassword(credentials.password);
    const storedHash = mockHashedPassword; // In production, retrieve from secure storage

    // Use timing-safe comparison to prevent timing attacks
    return timingSafeEqual(
      Buffer.from(mockHashedPassword),
      Buffer.from(storedHash)
    );
  }

  /**
   * Creates a security context for an authenticated session
   */
  private createSecurityContext(): SecurityContext {
    const now = new Date();
    return {
      issuedAt: now,
      expiresAt: new Date(now.getTime() + this.sessionDuration * 1000),
      permissions: ["read", "write"], // Would be based on user role in production
      tlsVersion: "TLSv1.3",
      cipherSuite: "TLS_AES_256_GCM_SHA384",
    };
  }

  /**
   * Generates a cryptographically secure session ID
   */
  private generateSessionId(): string {
    return randomBytes(32).toString("hex");
  }

  /**
   * Creates a failed authentication result
   */
  private createFailedResult(error: string): AuthResult {
    return {
      success: false,
      error,
    };
  }

  /**
   * Hashes a password using a secure algorithm
   * In production, use a proper password hashing library
   */
  private async hashPassword(password: string): Promise<string> {
    const salt = randomBytes(16).toString("hex");
    return createHash("sha256")
      .update(salt + password)
      .digest("hex");
  }
}
